\name{calculateRPM}
\alias{calculateRPM}
\alias{calculateRPM,r3Cseq-method}
\title{calculate read per million (RPM)}

\description{
  Normalize 3C-seq data by transforming raw reads to read per million per each restriction fragment 
}

\usage{
calculateRPM(object)
}


\arguments{

	\item{object}{
		r3Cseq object. The object might contain the raw read generated by function \link{getReadCountPerRestrictionFragment}
	}
}

\seealso{
  \link{getCoverage},
  \link{getReadCountPerRestrictionFragment},
  \link{contrRPM},
  \link{expRPM}
}

\author{
	S. Thongjuea
}

\examples{
 	
####Create the r3Cseq object#############
	library(BSgenome.Mmusculus.UCSC.mm9)
    load(system.file("data","example.data.rda",package="r3Cseq"))
	calculateRPM(my.data)
}
%\keyword{normalization}
