\name{contrReadCount}
\alias{contrReadCount}
\alias{contrReadCount,r3CseqCommon-method}
\title{get read count  per region for the control}

\description{
  get the read count per region for the control
}

\usage{
contrReadCount(object)
}


\arguments{

	\item{object}{
		r3Cseq object
	}
}

\seealso{
  \link{expReadCount},
  \link{getReadCountPerRestrictionFragment}
}

\author{
	S. Thongjuea
}

\examples{

#See the vignette	
}
%\keyword{get reads}
