\name{export3Cseq2bedGraph}
\alias{export3Cseq2bedGraph}
\alias{export3Cseq2bedGraph,r3Cseq-method}

\title{export interaction regions to the 'bedGraph' format}

\description{
  export interaction regions from RagedData to the bedGraph format, which
  suitable for uploading to the UCSC genome browser
}

\usage{
export3Cseq2bedGraph(object,datatype=c("rpm","raw_read"))
}


\arguments{

	\item{object}{
		r3Cseq object, The object might contain the interaction regions generated by function \link{getInteractions}
	}
	
	\item{datatype}{
			raw_read : read count per restriction fragment
			rpm : read per million per restriction fragment
	}
}

\value{
  The text file in 'bedGraph' format
}

\seealso{

  \link{getInteractions}
}

\author{
	S. Thongjuea
}

\examples{
 	
 ####Create the r3Cseq object#############
	library(BSgenome.Mmusculus.UCSC.mm9)
    load(system.file("data","example.data.rda",package="r3Cseq"))
	calculateRPM(my.data)
	getInteractions(my.data)
	export3Cseq2bedGraph(my.data,datatype="rpm")	
}
%\keyword{}
