\name{exportBatchInteractions2text}
\alias{exportBatchInteractions2text}
\alias{exportBatchInteractions2text,r3CseqInBatch-method}

\title{export identified interaction regions to the tab separated format for replicates analysis}

\description{
  export interaction regions from RagedData to the tab separated format for replicates analysis
}

\usage{
exportBatchInteractions2text(object)
}


\arguments{

	\item{object}{
		r3CseqInBatch object
	}
}

\value{
  The text file in the tab separated format
}

\seealso{

  \link{export3Cseq2bedGraph},
  \link{exportInteractions2text}
}

\author{
	S. Thongjuea
}

\examples{
 #See the vignette		
}
%\keyword{export identified interaction regions}
