\name{exportInteractions2text}
\alias{exportInteractions2text}
\alias{exportInteractions2text,r3Cseq-method}

\title{export interaction regions to the tab separated format}

\description{
  export interaction regions from RagedData to the tab separated format
}

\usage{
exportInteractions2text(object)
}


\arguments{

	\item{object}{
		r3Cseq object, The object might contain the interaction regions generated by function \link{getInteractions}
	}
}

\value{
  The text file in the tab separated format
}

\seealso{

  \link{getInteractions}
}

\author{
	S. Thongjuea
}

\examples{
 	
 ####Create the r3Cseq object#############
	library(BSgenome.Mmusculus.UCSC.mm9)
    load(system.file("data","example.data.rda",package="r3Cseq"))
	calculateRPM(my.data)
	getInteractions(my.data)
	exportInteractions2text(my.data)	
}
%\keyword{}
