\name{expReadCount}
\alias{expReadCount}
\alias{expReadCount,r3CseqCommon-method}
\title{get read count  per region for the experiment}

\description{
  get the read count per region for the experiment
}

\usage{
expReadCount(object)
}


\arguments{

	\item{object}{
		r3Cseq
	}
}

\seealso{
  \link{contrReadCount},
   \link{getReadCountPerRestrictionFragment}
}

\author{
	S. Thongjuea
}

\examples{

#See the vignette	
}
%\keyword{get reads}
