\name{expRPM}
\alias{expRPM}
\alias{expRPM,r3Cseq-method}
\title{get read per million (RPM) for the experiment}

\description{
  get the normalized 3C-seq data (RPM) for the experiment
}

\usage{
expRPM(object)
}


\arguments{

	\item{object}{
		r3Cseq object. The object might contain the raw read generated by function \link{getReadCountPerRestrictionFragment}
	}
}

\seealso{
  \link{calculateRPM},
  \link{contrRPM}
}

\author{
	S. Thongjuea
}

\examples{
 	
####Create the r3Cseq object#############
	library(BSgenome.Mmusculus.UCSC.mm9)
    load(system.file("data","example.data.rda",package="r3Cseq"))
	calculateRPM(my.data)
	experiment.rpm <-expRPM(my.data)
}
%\keyword{normalization}
