\name{generate3CseqReport}
\alias{generate3CseqReport}
\alias{generate3CseqReport,r3Cseq-method}
\title{generate reports for analysis results from r3Cseq}

\description{
  generate reports for analysis results from r3Cseq, the report contains 
  all plots in one pdf file and a text separated out put file.
}

\usage{
generate3CseqReport(object)
}


\arguments{

	\item{object}{
		r3Cseq object, The object might contain the interaction regions generated by function \link{getInteractions}
	}
}

\value{
  The text file in the tab separated format and the pdf file of all plots
}

\seealso{

  \link{getInteractions},
  \link{exportInteractions2text}
  \link{plotOverviewInteractions}, 
  \link{plotInteractionsPerChromosome}, 
  \link{plotInteractionsNearViewpoint},
  \link{plot3Cecdf}
}

\author{
	S. Thongjuea
}

\examples{
 	
 ####Create the r3Cseq object#############
	library(BSgenome.Mmusculus.UCSC.mm9)
    load(system.file("data","example.data.rda",package="r3Cseq"))
	calculateRPM(my.data)
	getInteractions(my.data)
	generate3CseqReport(my.data)
}
%\keyword{}
