\name{Myb_prom_FL}
\alias{Myb_prom_FL}
\title{Myb_prom_FL a data set for the example of r3Cseq analysis}
\description{
The example aligned reads generated by 3C-Seq protocol from fetal liver. The promoter region of the Myb's gene was selected as the viewpoint.
This data was transformed from aligned reads shown in the BAM file to GRanged object by using Rsamtools.  
}
\keyword{datasets}