\name{contrInteractionRegions}
\alias{contrInteractionRegions}
\alias{contrInteractionRegions,r3CseqCommon-method}
\title{get interaction regions from the control}

\description{
  get all identified interaction regions from the control
}

\usage{
contrInteractionRegions(object)
}


\arguments{

	\item{object}{
		r3Cseq or r3CseqInBatch object
	}
}

\value{
   The candidate interaction regions show in the IRange object
}

\seealso{
  \link{expInteractionRegions},
  \link{getInteractions}
}

\author{
	S. Thongjuea
}

\examples{
 	
#See the vignette
}
%\keyword{}
