\name{contrRPM}
\alias{contrRPM}
\alias{contrRPM,r3CseqCommon-method}
\title{get read per million (RPM) for the control}

\description{
  get the normalized 3C-seq data (RPM) for the control
}

\usage{
contrRPM(object)
}


\arguments{

	\item{object}{
		r3Cseq or r3CseqInBatch object
	}
}

\seealso{
  \link{calculateRPM},
  \link{expRPM}
}

\author{
	S. Thongjuea
}

\examples{
 	#See the vignette
}
%\keyword{normalization}
