\name{expInteractionRegions}
\alias{expInteractionRegions}
\alias{expInteractionRegions,r3CseqCommon-method}
\title{get interaction regions from the experiment}

\description{
  get identified interaction regions from the experiment
}

\usage{
expInteractionRegions(object)
}


\arguments{

	\item{object}{
		r3Cseq or r3CseqInBatch object
	}
}

\value{
   The candidate interaction regions show in the IRange object
}

\seealso{

  \link{getInteractions},
  \link{contrInteractionRegions}
}

\author{
	S. Thongjuea
}

\examples{

#See the vignette	
}
%\keyword{}
