\name{export3Cseq2bedGraph}
\alias{export3Cseq2bedGraph}
\alias{export3Cseq2bedGraph,r3Cseq-method}

\title{export interaction regions to the 'bedGraph' format}

\description{
  export interaction regions from RagedData to the bedGraph format, which
  suitable for uploading to the UCSC genome browser
}

\usage{
export3Cseq2bedGraph(object,datatype=c("rpm","read_count"))
}


\arguments{

	\item{object}{
		r3Cseq object, The object might contain the interaction regions generated by function \link{getInteractions}
	}
	
	\item{datatype}{
			read_count : read count per restriction fragment
			rpm : normalized read per million per restriction fragment
	}
}

\value{
  The text file in 'bedGraph' format
}

\seealso{

  \link{exportInteractions2text}
}

\author{
	S. Thongjuea
}

\examples{
#See the vignette	
}
%\keyword{export identified interaction regions}
