\name{exportInteractions2text}
\alias{exportInteractions2text}
\alias{exportInteractions2text,r3Cseq-method}

\title{export identified interaction regions to the tab separated format}

\description{
  export interaction regions from RagedData to the tab separated format
}

\usage{
exportInteractions2text(object)
}


\arguments{

	\item{object}{
		r3Cseq object
	}
}

\value{
  The text file in the tab separated format
}

\seealso{

  \link{export3Cseq2bedGraph}
}

\author{
	S. Thongjuea
}

\examples{
 #See the vignette		
}
%\keyword{export identified interaction regions}
