\name{expRPM}
\alias{expRPM}
\alias{expRPM,r3CseqCommon-method}
\title{get read per million (RPM) for the experiment}

\description{
  get the normalized 3C-seq data (RPM) for the experiment
}

\usage{
expRPM(object)
}


\arguments{

	\item{object}{
		r3Cseq or r3CseqInBatch
	}
}

\seealso{
  \link{calculateRPM},
  \link{contrRPM}
}

\author{
	S. Thongjuea
}

\examples{

#See the vignette	
}
%\keyword{normalization}
