\name{getInteractions}
\alias{getInteractions}
\alias{getInteractions,r3Cseq-method}
\title{assign p-value and fold change to candidate interaction regions}

\description{
  Assign p-value and fold change to each candidate interaction regions by using
  empirical distribution function
}

\usage{
getInteractions(object)
}


\arguments{

	\item{object}{
		r3Cseq object. The object might contain the RPM generated by function \link{getReadCountPerRestrictionFragment}
		following by \link{calculateRPM}
	}
}

\value{
   The candidate interaction regions show in the RangedData
}

\seealso{
  \link{getCoverage},
  \link{getReadCountPerRestrictionFragment},
  \link{calculateRPM}
}



\author{
	S. Thongjuea
}

\examples{
 	
####Create the r3Cseq object#############
	library(BSgenome.Mmusculus.UCSC.mm9)
	load(system.file("data","example.data.rda",package="r3Cseq"))
	calculateRPM(my.data)
	getInteractions(my.data)
}
%\keyword{}
